pyproject.toml
[project]
name = "hack-prep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "anthropic>=0.36.0",
    "boto3>=1.35.39",
    "elevenlabs>=1.9.0",
    "instructor>=1.5.2",
    "ipykernel>=6.29.5",
    "llama-index-utils-workflow>=0.2.1",
    "llama-index>=0.11.17",
    "lumaai>=1.1.0",
    "moviepy>=1.0.3",
    "openai-whisper>=20240930",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "reflex>=0.6.2.post1",
    "replicate==0.34.2",
    "requests>=2.32.3",
    "speechrecognition>=3.10.4",
]
